:tada:feat(models): add UnetGenerator model
The UnetGenerator model has been added to the `models.ts` file. This model implements the U-Net architecture for image segmentation using PaddlePaddle. It consists of several downsampling and upsampling layers, as well as a center layer that concatenates the outputs of the downsampling layers.

The `Downsample` and `Upsample` classes are also included in the `models.ts` file to support the implementation of the UnetGenerator. Additionally, a discriminator model, `NLayerDiscriminator`, has been defined.